+++ Sudoku.Shared/EmptySolver.cs
@@ -6,67 +6,9 @@ namespace Sudoku.Shared
 {
     public class EmptySolver : ISudokuSolver
     {
-        int population_size = 1000;
-
         public SudokuGrid Solve(SudokuGrid s)
         {
-            var cleanSudoku = new SudokuCleaning(s);
-            SudokuGrid sudoku = cleanSudoku.sudoku;
-            Dictionary<int, List<int>> mask = cleanSudoku.mask; 
-
-            var selection = new TournamentSelection();
-            var crossover = new UniformCrossover();
-            var mutation = new UniformMutation();
-            var fitness = new SudokuFitness();
-            var chromosome = new SudokuChromosome(sudoku, mask);
-            
-            Console.WriteLine("GA running...");
-            while (true)
-            {
-                var population = new Population(population_size, population_size, chromosome);
-                var ga = new GeneticAlgorithm(population, fitness, selection, crossover, mutation);
-                ga.Termination = new OrTermination(new FitnessStagnationTermination(20), new FitnessThresholdTermination(1));
-                // ga.OperatorsStrategy = new TplOperatorsStrategy();
-	            // ga.TaskExecutor = new TplTaskExecutor();
-
-                ga.GenerationRan += (sender, e) =>
-                {
-                    var bestFitness = ga.BestChromosome.Fitness;
-                    Console.WriteLine($"Generation {ga.GenerationsNumber}: Best Fitness = {bestFitness}");
-                };
-
-                ga.Start();
-
-                if (ga.BestChromosome.Fitness == 1)
-                    return ConvertChromosomeToSudokuGrid(ga.BestChromosome);
-
-                /* To put if we use Ilyas's version
-                {
-                    var sudokuChromosome = ga.BestChromosome as SudokuChromosome;
-                    return sudokuChromosome.getSolution();
-                }
-                */
-
-
-
-                population_size = (int)Math.Round(population_size * 1.5);
-            }
+            return s.CloneSudoku();
         }
-
-        public static SudokuGrid ConvertChromosomeToSudokuGrid(IChromosome bestChromosome)
-        {
-
-            int[,] sudoku = new int[9,9];
-            for (int i = 0; i < 9; i++) 
-            {
-                var gene = (int[])bestChromosome.GetGene(i).Value;
-
-                for (int j = 0; j < 9; j++)
-                    sudoku[i,j] = gene[j];
-            }
-
-            return new SudokuGrid { Cells = sudoku};
-        }
-
     }
 }
