+++ Sudoku.sln
@@ -28,6 +28,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sudoku.DancingLinks", "Sudo
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sudoku.GeneticAlgorithm", "Sudoku.GeneticAlgorithm\Sudoku.GeneticAlgorithm.csproj", "{DE02E4A8-ABF5-4249-9A5A-6C372DF9C56A}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sudoku.GeneticAlg", "Sudoku.GeneticAlg\Sudoku.GeneticAlg.csproj", "{ECE36C85-7FFF-4B1F-BF73-ADB280875E97}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -70,6 +72,10 @@ Global
 		{DE02E4A8-ABF5-4249-9A5A-6C372DF9C56A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DE02E4A8-ABF5-4249-9A5A-6C372DF9C56A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DE02E4A8-ABF5-4249-9A5A-6C372DF9C56A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{ECE36C85-7FFF-4B1F-BF73-ADB280875E97}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{ECE36C85-7FFF-4B1F-BF73-ADB280875E97}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{ECE36C85-7FFF-4B1F-BF73-ADB280875E97}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{ECE36C85-7FFF-4B1F-BF73-ADB280875E97}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
